requirements.txt
# MultiModal TaskGate — Requirements
# Install with: pip install -r requirements.txt

# ─────────────────────────────────────────────
# Core Deep Learning
# ─────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ─────────────────────────────────────────────
# Transformers & NLP
# ─────────────────────────────────────────────
transformers>=4.35.0
tokenizers>=0.15.0
datasets>=2.14.0

# ─────────────────────────────────────────────
# OCR
# ─────────────────────────────────────────────
easyocr>=1.7.0

# ─────────────────────────────────────────────
# Image Processing
# ─────────────────────────────────────────────
Pillow>=10.0.0
opencv-python>=4.8.0

# ─────────────────────────────────────────────
# Data Handling
# ─────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
jsonlines>=3.1.0
scikit-learn>=1.3.0

# ─────────────────────────────────────────────
# Training Utilities
# ─────────────────────────────────────────────
tqdm>=4.65.0
pyyaml>=6.0
accelerate>=0.24.0

# ─────────────────────────────────────────────
# Experiment Tracking
# ─────────────────────────────────────────────
wandb>=0.15.0

# ─────────────────────────────────────────────
# Visualization & Analysis
# ─────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ─────────────────────────────────────────────
# Demo App
# ─────────────────────────────────────────────
streamlit>=1.28.0

# ─────────────────────────────────────────────
# Statistical Testing
# ─────────────────────────────────────────────
scipy>=1.11.0